+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2017.
 * Improved admin interface for groups.
 * Added support for Yandex Translate API.
 * Improved speed of sitewide search.
+* Add project wide search.
 
 weblate 2.11
 ------------